+++ BookCat.Site/Repos/BookRepo.cs
@@ -6,12 +6,12 @@ namespace BookCat.Site.Repos;
 public class BookRepo : IRepo<Book>
 {
     private readonly AppDbContext _db;
-    private readonly IRepo<BookIdentifier> _bookIdfs;
+    private readonly IRepo<BookIdentifier> _identifiers;
 
     public BookRepo(AppDbContext appDbContext, IRepo<BookIdentifier> bookIdentifierRepo)
     {
         _db = appDbContext;
-        _bookIdfs = bookIdentifierRepo;
+        _identifiers = bookIdentifierRepo;
     }
     
     public async Task AddAsync(Book entity)
@@ -19,7 +19,7 @@ public class BookRepo : IRepo<Book>
         Book book = (await _db.Books.AddAsync(entity)).Entity;
         foreach (var item in book.Identifiers)
         {
-            await _bookIdfs.AddAsync(item);
+            await _identifiers.AddAsync(item);
         }
         
         await _db.SaveChangesAsync();
@@ -30,26 +30,30 @@ public class BookRepo : IRepo<Book>
         return await _db.Books.Include(b => b.Identifiers).Include( b => b.Reviews).Include(b => b.AddedBy).ToListAsync();
     }
 
-    public async Task<Book> GetByIdAsync(Guid id)
+    public async Task<Book?> GetByIdAsync(Guid id)
     {
-        Book? book = await _db.Books.FindAsync(id) ?? throw new KeyNotFoundException();
+        Book? book = await _db.Books.FindAsync(id);
+        if (book is null) return null;
         await _db.Entry(book).Collection(b => b.Identifiers).LoadAsync();
-        await _db.Entry(book).Collection(b => b.Reviews).LoadAsync();
         await _db.Entry(book).Reference(b => b.AddedBy).LoadAsync();
         return book;
     }
 
     public async Task UpdateAsync(Book entity)
     {
-        throw new NotImplementedException(nameof(BookRepo) + "UpdateAsync, never update books");
+        _db.Books.Update(entity);
+
+        await _db.SaveChangesAsync();
     }
 
-    public async Task DeleteAsync(Guid id)
+    public async Task<bool> DeleteAsync(Guid id)
     {
-        Book? book = await _db.Books.FindAsync(id) ?? throw new KeyNotFoundException();
+        Book? book = await _db.Books.FindAsync(id);
+        if (book is null) return false;
         _db.Books.Remove(book);
 
         await _db.SaveChangesAsync();
+        return true;
     }
 
     public async Task DeleteAsync(Book entity)
@@ -69,7 +73,7 @@ public class BookRepo : IRepo<Book>
         return _db.Books.Where(b => b.AddedById == id);
     }
 
-    public async Task<int> GetCount()
+    public async Task<int> GetCountAsync()
     {
         return await _db.Books.CountAsync();
     }
